Split ProductService into interface and implementation class
