+++ backend/spring-app/spring-app/src/main/java/com/avdo/spring/app/service/ProductService.java
@@ -12,46 +12,10 @@ import org.springframework.stereotype.Service;
 import java.util.List;
 
 @Service
-public class ProductService {
+public interface ProductService {
 
-    private final ProductRepository productRepository;
-    private final CategoryRepository categoryRepository;
-
-    @Autowired
-    public ProductService(
-            ProductRepository productRepository, CategoryRepository categoryRepository) {
-        this.productRepository = productRepository;
-        this.categoryRepository = categoryRepository;
-    }
-
-    public ProductEntity createProduct(CreateProductRequest createProductRequest) {
-        Category category = categoryRepository.findCategoryByName(createProductRequest.getCategory());
-        ProductEntity product = ProductMapper.mapToProduct(createProductRequest, CategoryEntity.fromCategory(category));
-        return productRepository.save(product);
-    }
-
-    public List<ProductEntity> findAllProducts() {
-        return productRepository.findAll();
-    }
-
-    public ProductEntity findById(Long id) {
-        return productRepository.findById(id).orElseThrow();
-    }
-
-    public List<ProductEntity> findSameCategoryProducts(String category) {
-        return productRepository.findByCategoryEntity_Name(category);
-    }
-
-    private static class ProductMapper {
-
-        private static ProductEntity mapToProduct(CreateProductRequest createProductRequest, CategoryEntity categoryEntity) {
-            ProductEntity product = new ProductEntity();
-            product.setName(createProductRequest.getName());
-            product.setPrice(createProductRequest.getPrice());
-            product.setStock(createProductRequest.getStock());
-            product.setDescription(createProductRequest.getDescription());
-            product.setCategoryEntity(categoryEntity);
-            return product;
-        }
-    }
+    ProductEntity createProduct(CreateProductRequest createProductRequest);
+    List<ProductEntity> findAllProducts();
+    public ProductEntity findById(Long id);
+    List<ProductEntity> findSameCategoryProducts(String category);
 }
